requirements.txt
# 
gspread
oauth2client
python-dotenv
sendgrid